refactor(ItemsPage): Move data acquisition logic from BestItemList to ItemsPage
