+++ src/components/BestItemList.js
@@ -1,39 +1,10 @@
-import { useEffect, useState } from "react";
-
-import { MOBILE, TABLET, PC } from "../utils/magicLiterals";
-import { getBestItems } from "../api";
+import Item from "./Item";
 import "./BestItemList.css";
 
-import Item from "./Item";
-
-const numItems = {
-  [MOBILE]: 1,
-  [TABLET]: 2,
-  [PC]: 4,
-};
-
-function BestItemList({ device }) {
-  const [bestItems, setBestItems] = useState([]);
-  const [bestItemsUsing, setBestItemsUsing] = useState([]);
-
-  const handleLoad = async () => {
-    const { list } = await getBestItems();
-    setBestItems(list);
-    setBestItemsUsing(list.slice(0, numItems[device]));
-  };
-
-  useEffect(() => {
-    handleLoad();
-  }, []);
-
-  useEffect(() => {
-    const newBestItemUsing = bestItems.slice(0, numItems[device]);
-    setBestItemsUsing(newBestItemUsing);
-  }, [device]);
-
+function BestItemList({ items }) {
   return (
     <ol className="item-list">
-      {bestItemsUsing.map((item) => (
+      {items.map((item) => (
         <li key={item.id}>
           <Item item={item} />
         </li>
